Implement caches
With this we add two caches: First one for invalid username/password
login attempts to prevent excessive requests to CAS for local SCM
users or requests with api tokens. Second, we configure a simple cache
to be used by shiro to reduce CAS requests in general.

Doing so, we also prevent throttling by CAS, because the CAS
only sees invalid requests and might block the SCM server in
such cases.